+++ Memoriae/Screen/Levels/TOMMPicture/TOMMPictureAnswersViewController.swift
@@ -11,7 +11,7 @@ import UIKit
 class TOMMPictureAnswersViewController: UIViewController {
     
     private var level: Level?
-    private var difficulty: Double?
+    private var difficulty: Int?
     private var targetNumber: Int?
     private var currentTarget: TOMMPictureSet?
 
@@ -152,7 +152,7 @@ class TOMMPictureAnswersViewController: UIViewController {
         answers?[index]
     }
 
-    func setTestDetails(level: Level?, difficulty: Double?, targetNumber: Int?, currentTarget: TOMMPictureSet?) {
+    func setTestDetails(level: Level?, difficulty: Int?, targetNumber: Int?, currentTarget: TOMMPictureSet?) {
 
         self.level = level
         self.difficulty = difficulty
